chapter3_4_aianimals/ab_test_proxy/pyproject.toml
[tool.poetry]
name = "ab_test_proxy"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.22.0"
fastapi = "^0.73.0"
gunicorn = "^20.1.0"
uvicorn = "^0.17.4"
pydantic = "^1.9.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
